+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 **Features**:
 
+* add connection timeout for SessionInfo #21, @flipback
 * add Identity Object #5, @flipback
 * add utils::LogAppenderIf to customise logging #5, @flipback
 * add support of LWORD type #6, @flipback